refactor: split acp manager task and timeout helpers
